Update
- Stok Transfer 🆕
- Stok Adjustment 🆕
- Stok Opname 🆕
- Report Stok Opanem 🆕
- Dashboard Inventory 🆕

+++ database/migrations/2025_07_13_143154_create_purchase_payments_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('purchase_payments')) {
+            return;
+        }
+        
         Schema::create('purchase_payments', function (Blueprint $table) {
             $table->id();
             $table->foreignId('purchase_id')->constrained('purchases')->onDelete('cascade');
